Extract shared histogram chart view
Shared for both normal and torque layers
